Implemented ally detection for main player unit.
Find allies in the vicinity of only the main player unit. Detects when player attacks enemy with units, and checks if there are ally units nearby the main player unit. This brings up the issue where if units are detected near the main player unit when it is far from the unit the player is actually using to attack, it will use those units as cohesion, when it should not be since it isn't the main player unit attacking.