Limit ally detection to surroundings of selected player unit
Fully implemented accurate detection of allies near the unit the player has selected. This can now be used to factor in cohesion bonuses in any action. For now, a message is shown in the console to reflect allies nearby the unit clicked, as the feature is not implemented in any gameplay yet.