Added 2 NPC's and some dialogue
I made it so that there are 2 NPC's. 1 that requires you to have finished the coins quest and 1 that you can just talk to right away for some random extra dialogue. I am still trying to figure out how the whole system works so this is more practice to get a grasp of the flow.